Added flags and corrected appropriate warnings

+++ done/test-file.c
@@ -37,7 +37,7 @@ int test(struct unix_filesystem *u) {
         if (error != 0) {
             return error;
         }
-        for (int i = 0; i < INODES_PER_SECTOR; i++) {
+        for (size_t i = 0; i < INODES_PER_SECTOR; i++) {
             struct inode in = sector[i];
             if (in.i_mode & IALLOC) {
                 print_sha_inode(u, in, ((s - u->s.s_inode_start) * INODES_PER_SECTOR) + i);
